refactor ticketinbox component to separete the into ticket component and implement new style for the skeleton
